+++ src/components/ApiKeyInputField/ApiKeyInputField.styled.tsx
@@ -1,11 +1,16 @@
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 import { InputFieldStyled } from "../../shared/StyledComponents";
 
 export const ApiKeyInputFieldStyled = styled(InputFieldStyled)<{
   isMissing: boolean;
 }>`
-  border: ${(p) =>
-    p.isMissing ? "3px solid red" : "1px solid hsl(0, 0%, 80%)"};
+  border: 1px solid hsl(0, 0%, 80%);
+
+  ${(p) =>
+    p.isMissing &&
+    css`
+      border: 3px solid red;
+    `}
 `;
 
 export const ApiKeyInputWrapper = styled.div`
