--- src/components/HolidayTypeFilter/HolidayTypeSelectField.styled.tsx
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const HolidayTypeSelectFieldWrapper = styled.div`
-  & #react-select-2-placeholder {
-    font-size: 14px;
-  }
-`;
