+++ src/components/HolidayTypeSelectField/HolidayTypeSelectField.constants.ts
@@ -1,0 +1,27 @@
+export const holidayTypeOptions = [
+  {
+    value: ["National holiday"],
+    label: "National",
+    color: "#79f2c0",
+  },
+  {
+    value: [
+      "Observance",
+      "Season",
+      "Clock change/Daylight Saving Time",
+      "Half-day holiday",
+    ],
+    label: "Observance",
+    color: "#b2d4ff",
+  },
+  {
+    value: ["Common local holiday", "Local holiday"],
+    label: "Local",
+    color: "#ffbdad",
+  },
+  {
+    value: ["Religious"],
+    label: "Religious",
+    color: "#c0b6f3",
+  },
+];
